requirements.txt
discord.py==0.16.12
beautifulsoup4==4.6.0
bs4==0.0.1
asyncio==3.4.3
aiohttp==3.4.4
Pillow==5.2.0
aiofiles==0.4.0